Refactor protocol disruption logic
Separate the generic protocol disruption logic
from protocol specific logic, implemented as a proxy.
